requirements.txt
markdownify==0.11.6
beautifulsoup4==4.9.3
requests==2.28.2
python-dotenv==0.21.1
